refactor(defaultComponent/parser): Move default component parser
In a new folder named defaultParser
