+++ Engine/Shared/DefaultParser/PlayerParser/PlayerParser.hpp
@@ -1,0 +1,20 @@
+/*
+** EPITECH PROJECT, 2024
+** PlayerParser
+** File description:
+** PlayerParser
+*/
+
+#pragma once
+
+#include "../../DefaultComponents/Entities/Player/PlayerComponent.hpp"
+
+#include <memory>
+
+/**
+ * @brief Parse a Player Json object.
+ *
+ * @param pathFile Path to the Json file.
+ * @return std::shared_ptr<PlayerComponent>
+ */
+std::shared_ptr<PlayerComponent> parsePlayer(std::string pathFile);
