+++ Engine/Shared/DefaultParser/Position2DParser/Position2DParser.hpp
@@ -1,0 +1,20 @@
+/*
+** EPITECH PROJECT, 2024
+** Position2DParser
+** File description:
+** Position2DParser
+*/
+
+#pragma once
+
+#include "../../DefaultComponents/Position2D/Position2DComponent.hpp"
+
+#include <memory>
+
+/**
+ * @brief Parse a Position2D Json object.
+ *
+ * @param pathFile Path to the Json file.
+ * @return std::shared_ptr<Position2DComponent>
+ */
+std::shared_ptr<Position2DComponent> parsePosition2D(std::string pathFile);
