+++ Engine/Shared/DefaultParser/ScaleParser/ScaleParser.hpp
@@ -1,0 +1,20 @@
+/*
+** EPITECH PROJECT, 2024
+** ScaleParser
+** File description:
+** ScaleParser
+*/
+
+#pragma once
+
+#include "../../DefaultComponents/Scale/ScaleComponent.hpp"
+
+#include <memory>
+
+/**
+ * @brief Parse a Scale Json object.
+ *
+ * @param pathFile Path to the Json file.
+ * @return std::shared_ptr<ScaleComponent>
+ */
+std::shared_ptr<ScaleComponent> parseScale(std::string pathFile);
