+++ Engine/Shared/DefaultParser/TextureRectParser/TextureRectParser.hpp
@@ -1,0 +1,20 @@
+/*
+** EPITECH PROJECT, 2024
+** TextureRectParser
+** File description:
+** TextureRectParser
+*/
+
+#pragma once
+
+#include "../../DefaultComponents/TextureRect/TextureRectComponent.hpp"
+
+#include <memory>
+
+/**
+ * @brief Parse a TextureRect Json object.
+ *
+ * @param pathFile Path to the Json file.
+ * @return std::shared_ptr<TextureRectComponent>
+ */
+std::shared_ptr<TextureRectComponent> parseTextureRect(std::string pathFile);
